+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Updated StencilSwiftKit to 1.0.2 which includes Stencil 0.9.0
 - Adding docset to release archive
 - Moved to CocoaPods 1.2.1
+- Made Array.parallelMap's block non-escaping
 
 
 ## 0.6.0